requirements.txt
# Web framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Core scientific stack
numpy==2.0.2
pandas==2.2.3

# ML stack
scikit-learn==1.5.2
xgboost==2.1.4

# Database driver (adjust if you use something else)
psycopg2-binary==2.9.9  # for PostgreSQL

# Optional: gunicorn if you deploy on Linux server
gunicorn==23.0.0
